requirements.txt
colorama
numpy
Pillow
matplotlib
scipy
torch
torchvision